+++ authentikos/go.mod
@@ -6,8 +6,13 @@ require (
 	github.com/Masterminds/goutils v1.1.0 // indirect
 	github.com/Masterminds/semver v1.5.0 // indirect
 	github.com/Masterminds/sprig v2.22.0+incompatible
+	github.com/bradleyfalzon/ghinstallation v1.1.1
+	github.com/evanphx/json-patch v4.5.0+incompatible // indirect
+	github.com/google/go-cmp v0.3.1
 	github.com/huandu/xstrings v1.3.0 // indirect
+	github.com/mattbaird/jsonpatch v0.0.0-20200820163806-098863c1fc24
 	github.com/mitchellh/copystructure v1.0.0 // indirect
+	github.com/pkg/errors v0.8.1 // indirect
 	github.com/spf13/pflag v1.0.3
 	golang.org/x/crypto v0.0.0-20190701094942-4def268fd1a4 // indirect
 	golang.org/x/oauth2 v0.0.0-20190604053449-0f29369cfe45
@@ -16,4 +21,5 @@ require (
 	k8s.io/api v0.0.0-20191005115622-2e41325d9e4b
 	k8s.io/apimachinery v0.0.0-20191006235458-f9f2f3f8ab02
 	k8s.io/client-go v0.0.0-20191008115822-1210218b4a26
+	k8s.io/klog v1.0.0
 )
