+++ authentikos/pkg/plugins/google/BUILD.bazel
@@ -1,0 +1,32 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["google.go"],
+    importpath = "istio.io/test-infra/authentikos/pkg/plugins/google",
+    visibility = ["//visibility:public"],
+    deps = [
+        "@com_github_masterminds_sprig//:go_default_library",
+        "@com_github_spf13_pflag//:go_default_library",
+        "@io_k8s_api//core/v1:go_default_library",
+        "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
+        "@io_k8s_client_go//plugin/pkg/client/auth:go_default_library",
+        "@io_k8s_client_go//rest:go_default_library",
+        "@io_k8s_client_go//tools/clientcmd:go_default_library",
+        "@io_k8s_klog//:go_default_library",
+        "@org_golang_google_api//option:go_default_library",
+        "@org_golang_google_api//transport:go_default_library",
+        "@org_golang_x_oauth2//:go_default_library",
+        "@org_golang_x_oauth2//google:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = ["google_test.go"],
+    embed = [":go_default_library"],
+    deps = [
+        "@com_github_spf13_pflag//:go_default_library",
+        "@org_golang_x_oauth2//:go_default_library",
+    ],
+)
